+++ app/src/main/java/org/simple/clinic/summary/PatientSummaryScreen.kt
@@ -23,7 +23,6 @@ import io.reactivex.ObservableTransformer
 import io.reactivex.disposables.CompositeDisposable
 import io.reactivex.disposables.Disposable
 import io.reactivex.rxkotlin.cast
-import io.reactivex.rxkotlin.ofType
 import io.reactivex.subjects.PublishSubject
 import kotlinx.android.parcel.Parcelize
 import org.simple.clinic.R
@@ -34,13 +33,16 @@ import org.simple.clinic.di.injector
 import org.simple.clinic.editpatient.EditPatientScreenKey
 import org.simple.clinic.facility.Facility
 import org.simple.clinic.facility.alertchange.AlertFacilityChangeSheet
-import org.simple.clinic.facility.alertchange.Continuation.ContinueToActivity
+import org.simple.clinic.facility.alertchange.Continuation.ContinueToScreenExpectingResult
 import org.simple.clinic.facility.alertchange.Continuation.ContinueToScreen_Old
 import org.simple.clinic.home.HomeScreenKey
 import org.simple.clinic.mobius.DeferredEventSource
 import org.simple.clinic.mobius.ViewRenderer
+import org.simple.clinic.navigation.v2.ExpectsResult
 import org.simple.clinic.navigation.v2.HandlesBack
 import org.simple.clinic.navigation.v2.Router
+import org.simple.clinic.navigation.v2.ScreenResult
+import org.simple.clinic.navigation.v2.Succeeded
 import org.simple.clinic.navigation.v2.fragments.BaseScreen
 import org.simple.clinic.patient.DateOfBirth
 import org.simple.clinic.patient.Gender
@@ -50,7 +52,6 @@ import org.simple.clinic.patient.businessid.BusinessId
 import org.simple.clinic.patient.businessid.Identifier
 import org.simple.clinic.patient.displayLetterRes
 import org.simple.clinic.router.ScreenResultBus
-import org.simple.clinic.router.screen.ActivityResult
 import org.simple.clinic.scheduleappointment.ScheduleAppointmentSheet
 import org.simple.clinic.summary.addphone.AddPhoneNumberDialog
 import org.simple.clinic.summary.linkId.LinkIdWithPatientSheet.LinkIdWithPatientSheetKey
@@ -59,7 +60,6 @@ import org.simple.clinic.summary.teleconsultation.messagebuilder.LongTeleconsult
 import org.simple.clinic.summary.updatephone.UpdatePhoneNumberDialog
 import org.simple.clinic.teleconsultlog.teleconsultrecord.screen.TeleconsultRecordScreenKey
 import org.simple.clinic.util.UserClock
-import org.simple.clinic.util.extractSuccessful
 import org.simple.clinic.util.messagesender.WhatsAppMessageSender
 import org.simple.clinic.util.toLocalDateAtZone
 import org.simple.clinic.widgets.UiEvent
@@ -81,7 +81,8 @@ class PatientSummaryScreen :
     PatientSummaryScreenUi,
     PatientSummaryUiActions,
     PatientSummaryChildView,
-    HandlesBack {
+    HandlesBack,
+    ExpectsResult {
 
   private val rootLayout
     get() = binding.rootLayout
@@ -244,10 +245,14 @@ class PatientSummaryScreen :
     // Not sure why but the keyboard stays visible when coming from search.
     rootLayout.hideKeyboard()
 
-    subscriptions.addAll(
-        setupChildViewVisibility(),
-        appointmentScheduleSheetClosed()
-    )
+    subscriptions.add(setupChildViewVisibility())
+  }
+
+  override fun onScreenResult(requestType: Parcelable, result: ScreenResult) {
+    if (requestType == ScreenRequest.ScheduleAppointmentSheet && result is Succeeded) {
+      val sheetOpenedFrom = ScheduleAppointmentSheet.sheetOpenedFrom(result)
+      appointmentScheduleSheetClosed.notify(ScheduledAppointment(sheetOpenedFrom))
+    }
   }
 
   @SuppressLint("CheckResult")
@@ -321,16 +326,6 @@ class PatientSummaryScreen :
       emitter.setCancellable { bloodPressureSummaryView.bpRecorded = null }
     }
   }
-
-  private fun appointmentScheduleSheetClosed() = screenResults
-      .streamResults()
-      .ofType<ActivityResult>()
-      .extractSuccessful(SUMMARY_REQCODE_SCHEDULE_APPOINTMENT) { intent ->
-        ScheduleAppointmentSheet.readExtra<ScheduleAppointmentSheetExtra>(intent)
-      }
-      .subscribe {
-        appointmentScheduleSheetClosed.notify(ScheduledAppointment(it!!.sheetOpenedFrom))
-      }
 
   private fun phoneNumberClicks(): Observable<UiEvent> {
     return contactTextView.clicks().map { ContactPatientClicked }
@@ -440,11 +435,12 @@ class PatientSummaryScreen :
       sheetOpenedFrom: AppointmentSheetOpenedFrom,
       currentFacility: Facility
   ) {
-    val scheduleAppointmentIntent = ScheduleAppointmentSheet.intent(requireContext(), patientUuid, ScheduleAppointmentSheetExtra(sheetOpenedFrom))
-
     router.push(AlertFacilityChangeSheet.Key(
         currentFacilityName = currentFacility.name,
-        continuation = ContinueToActivity(scheduleAppointmentIntent, SUMMARY_REQCODE_SCHEDULE_APPOINTMENT)
+        continuation = ContinueToScreenExpectingResult(
+            requestType = ScreenRequest.ScheduleAppointmentSheet,
+            screenKey = ScheduleAppointmentSheet.Key(patientUuid, sheetOpenedFrom)
+        )
     ))
   }
 
@@ -544,9 +540,10 @@ class PatientSummaryScreen :
   interface Injector {
     fun inject(target: PatientSummaryScreen)
   }
-}
 
-@Parcelize
-private data class ScheduleAppointmentSheetExtra(
-    val sheetOpenedFrom: AppointmentSheetOpenedFrom
-) : Parcelable
+  sealed class ScreenRequest : Parcelable {
+
+    @Parcelize
+    object ScheduleAppointmentSheet : ScreenRequest()
+  }
+}